Completed Jumping On Clouds.
Essentially this one is solved by using a lookahead pointer. To minimize the number of jumps, we always want to prioritize making the +2 jump. We only resort to a +1 jump when the +2 jump is an illegal move. We then use a counter to increment the number of jumps we've taken.